Mojarra defaultResourceMaxAge configuration property as long

+++ joinfaces-autoconfigure/src/main/java/org/joinfaces/autoconfigure/mojarra/MojarraProperties.java
@@ -268,7 +268,7 @@ public class MojarraProperties implements ServletContextInitParameterProperties 
 	 * the default value of 604800000 is 7 days).
 	 */
 	@ServletContextInitParameter(PREFIX + "defaultResourceMaxAge")
-	private Integer defaultResourceMaxAge;
+	private Long defaultResourceMaxAge;
 
 	/**
 	 * See issue 3684 for details.
